Add files via upload
These images are the results after the input thresholded image (see in attributes folder) is Bitwise ANDED with each of the thresholded stock images. Since a white pixel is denoted as a '1' and a black pixel is denoted as a '0', this is essentially overlapping both the input and stock image together and any white pixel that does not overlap with one another is blacked out.